remove roles, add permissions, change Int to Long in ids

+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Captcha.kt
@@ -1,8 +1,8 @@
 package io.github.e1turin.itmochan.entity
 
 data class Captcha(
-    val capthchaId : Int,
+    val capthchaId : Long,
     val answer : String?,
-    val pictureId : Int,
+    val pictureId : Long,
 )
