+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Comment.kt
@@ -3,25 +3,25 @@ package io.github.e1turin.itmochan.entity
 import java.time.LocalDateTime
 
 data class Comment(
-    val commentId : Int,
-    val threadId : Int,
+    val commentId : Long,
+    val threadId : Long,
     val title : String?,
     val content : String,
-    val userId : Int,
-    val reactionId : Int,
+    val userId : Long,
+    val reactionId : Long,
     val creationDate : LocalDateTime,
     val trashed : Boolean,
     val deleted : Boolean,
 )
 
 data class Reply(
-    val commentId : Int,
-    val replyCommentId : Int,
+    val commentId : Long,
+    val replyCommentId : Long,
 )
 
 data class CommentDTO(
-    val threadId : Int,
+    val threadId : Long,
     val title : String?,
     val content : String,
-    val userId : Int,
+    val userId : Long,
 )
