+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Reactions.kt
@@ -1,12 +1,12 @@
 package io.github.e1turin.itmochan.entity
 
 data class Reactions(
-    val rSetId : Int,
+    val rSetId : Long,
     val reactions : String,
 )
 
 data class ReactionDTO(
-    val rSetId : Int,
+    val rSetId : Long,
     val reaction : String,
 )
