+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Thread.kt
@@ -1,12 +1,12 @@
 package io.github.e1turin.itmochan.entity
 
 data class Thread(
-    val threadId : Int,
-    val topicId : Int,
-    val initCommentId : Int,
+    val threadId : Long,
+    val topicId : Long,
+    val initCommentId : Long,
     val popularity : Int,
 )
 
 data class ThreadDTO(
-    val topicId : Int,
+    val topicId : Long,
 )
